Cleaned GUI and minor refactoring
Moved check in errors to a new label
Fixed multiple label issues
- Multiline labels are now used where required
- Error labels now provide clear, meaningful details
Changed many comments
Added Javadoc comments to public methods
Added clearDetails, used to clear all customer-specific information from the gui. Called when an error is thrown on the check in screen, and when returning to the check in screen from elsewhere.