Swing error fix, and Customer.setCheckedIn() called in Flight.addCustomer()
QueueDisplay.onNotify() now calls SwingUtilities.invokeLater() on the updateDisplay() method, rather than updating directly. This should make the correct thread update the GUI.

Flights now set Customer to being checked in when the Customer is added to the flight, rather than Customer.setCheckedIn() needing to be explicitly called. Removed this call in Desk, and added error catch for AlreadyCheckedInException in Simulator.